Relocate dao related enum
